edits

+++ CANopen/motor/commands.c
@@ -53,13 +53,13 @@ int sort_read(int32_t* pos, my_can_frame f, int * num_of_reads)
 	switch(f.id) {
 		case(PDO_TX2_ID + 1):
 			enc = ((uint32_t)f.data[0]<<0) | ((uint32_t)f.data[1]<<8) | ((uint32_t)f.data[2]<<16) | ((uint32_t)f.data[3]<<24);
-			pos[0] = enc;//motor_enc_to_mm(enc);
+			pos[0] = (double)enc;//motor_enc_to_mm(enc);
 			*num_of_reads = *num_of_reads + 1;
 			printf("Node 1: %d\n", enc);
 			break;
 		case(PDO_TX2_ID + 2):
 			enc = ((uint32_t)f.data[0]<<0) | ((uint32_t)f.data[1]<<8) | ((uint32_t)f.data[2]<<16) | ((uint32_t)f.data[3]<<24);
-			pos[1] = enc;//motor_enc_to_mm(-enc);
+			pos[1] = (double)enc;//motor_enc_to_mm(-enc);
 			*num_of_reads = *num_of_reads + 1;
 			printf("Node 2: %d\n", enc);
 			break;
